third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-contain/container-queries/font-relative-units-dynamic-expected.txt

This is a testharness.js-based test.
[FAIL] cap units respond to changes
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] rcap units respond to changes
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
